+++ skywalking.c
@@ -248,6 +248,7 @@ ZEND_API void sky_execute_internal(zend_execute_data *execute_data, zval *return
             zf->common.scope->name) : NULL;
     const char *function_name = zf->common.function_name == NULL ? NULL : ZSTR_VAL(zf->common.function_name);
 
+    int is_procedural_mysqli = 0; // "Procedural style" or "Object oriented style" ?
     char *operationName = NULL;
     char *component = NULL;
     if (class_name != NULL) {
@@ -272,8 +273,29 @@ ZEND_API void sky_execute_internal(zend_execute_data *execute_data, zval *return
                 strcat(operationName, "->");
                 strcat(operationName, function_name);
             }
+        } else if (strcmp(class_name, "mysqli") == 0) {
+            if (strcmp(function_name, "query") == 0) {
+                component = (char *) emalloc(strlen("mysqli") + 1);
+                strcpy(component, "mysqli");
+                operationName = (char *) emalloc(strlen(class_name) + strlen(function_name) + 3);
+                strcpy(operationName, class_name);
+                strcat(operationName, "->");
+                strcat(operationName, function_name);
+            }
         }
     } else if (function_name != NULL) {
+        if (strcmp(function_name, "mysqli_query") == 0) {
+            class_name = "mysqli";
+            function_name = "query";
+            component = (char *) emalloc(strlen(class_name) + 1);
+            strcpy(component, class_name);
+            operationName = (char *) emalloc(strlen(class_name) + strlen(function_name) + 3);
+            strcpy(operationName, class_name);
+            strcat(operationName, "->");
+            strcat(operationName, function_name);
+
+            is_procedural_mysqli = 1;
+        }
     }
 
     if (operationName != NULL) {
@@ -323,7 +345,24 @@ ZEND_API void sky_execute_internal(zend_execute_data *execute_data, zval *return
                     add_assoc_string(&tags, "db.data_source", (char *) stmt->dbh->data_source);
                 }
             }
+        } else if (function_name != NULL) {
+            if (strcmp(class_name, "mysqli") == 0 && strcmp(function_name, "query") == 0) {
+                add_assoc_string(&tags, "db.type", "mysqli");
+                // params
+                uint32_t arg_count = ZEND_CALL_NUM_ARGS(execute_data);
+                if (arg_count) {
+                    zval *p = is_procedural_mysqli ? ZEND_CALL_ARG(execute_data, 2) : ZEND_CALL_ARG(execute_data, 1);
+                    //db.statement
+                    switch (Z_TYPE_P(p)) {
+                        case IS_STRING:
+                            add_assoc_string(&tags, "db.statement", Z_STRVAL_P(p));
+                            break;
+
+                    }
+                }
+            }
         }
+
 
         zval temp;
         zval *spans = NULL;
@@ -1303,9 +1342,11 @@ PHP_MINIT_FUNCTION (skywalking) {
             return SUCCESS;
         }
 
+        // 用户自定义函数执行器(php脚本定义的类、函数)
         ori_execute_ex = zend_execute_ex;
         zend_execute_ex = sky_execute_ex;
 
+        // 内部函数执行器(c语言定义的类、函数)
         ori_execute_internal = zend_execute_internal;
         zend_execute_internal = sky_execute_internal;